fix empty QucsHomeDir from QSettings
On a brand new Debian installation the settings.contains("QucsHomeDir")
passes an empty string into QucsSettings.QucsHomeDir.

It cannot cd into the directory to populate the list of projects.
The list will be empty, crashing the pop_front() method.

Added a test to avoid loading from settings in case it is an empty string.