deploy constans.vams and disciplines.vams includes
Use a known include for Verilog-A models.
If these files are not passed to admsXml (-I path) it will create
new files on the working directory.
There is BUG in adms with the P_U0 define. It cannot parse it if used
in a model.
We use the floating point number instead of the:
`define P_U0 (4.0e-7 * `M_PI)

 * autotools and cmake deploy the .vams files
 * va2cpp now include the installed .vams files
 * make clean updated to clean up all the generated files