Corrected font metrics computation.
In Qt4 the actual font size depends on the device that will actually
paint the font. To avoid to explicitly ask the painter to compute
the font size we can use QFontMetrics::QFontMetrics(font, 0) to
always obtain the correct screen metric (behavior similar to the Qt3
default). All istances of QFontMetrics have now been so modified.