Fixed errors due to numerical precision.
Initial state is placed further away than 'duration' to be sure
that the switching is considered complete even in case of numerical
errors.
Calculation or the actual 'r' rearranged to reduce the errors due
to the limited machine precision.
Added asserts to check actual computed resistance vs.Ron and Roff.