fix crash when close file with editText focus
The repeat steps:
1. Open a schematic file.
2. Focus on a component property, show the property edit
3. Close the file
4. The program will crash

The source:
This bug is appeared since commit eadd7b

The cause:
When click on property edit, this line
qucs_action.cpp:1221, editText->reparent(...
will change the editText to child of the open document
At this time, if the document is closed, the editText will
be deleted by Qt
After that, qucs try to access to editText will access to invalid
memory.

The fix:
To be general, now editText will child of main app every time the
slotHideEdit() called, it's a little overhead, but guarantee safe.
function slotHideEdit() {
  ...
  editText->setParent(this, 0);
}