Change search-path messagebox to status bar message
Using "Save All" or "Save as..." displayed a modal message box
informing the user that "The schematic file path has been refreshed".
This message box is likely a test message remained in the main code
after the subcircuit search path feature has been implemented.
The search path refresh does not require the user immediate attention
so the modal messagebox has been replace with a simple message on
the statusbar.