don't try to recycle abused Graph instances.
(this fix has already been proposed in the dc-bias-points-bug
discussion, however, it did not find its way)

the suspected reason for the Graph abuse is the missing (standalone)
.dat file parser. Otherwise, no graphs seem to be involved in the bias
points afaics.

hopefully this can be amended later on, and will look close to something
like this:

biasPointContainer* b = new biasPointContainer(); // whatever this is.
SimOutputBiasPoints* d = somehowgettheresultsfromthedcpointsimulation();
b.attach(d); // keep track of it.

b->showsomedialog();
Schematic->attachbiaspoints(b); // display and keep track of it ...