Abort if there is no simulation component
The simulator is always started, regardless if there is no simulation
block on the schematic. The simulator returns a unclear warning about
'actions'.

This commit checks for the presence of at least one simulation component
prior to executing the simulator.

It also opens a information dialog to inform the user about necessary
simulation components.

This closes issue #259.