Correct Components Tab behavior when searching.
When using the Component Search in the Components Tab the component
category shown in the top combo box was not changed, but still showed
the last used category. Now when searching, the category text shown is
changed to "Search results" and restored to the last used category
when exiting the search.