[qucsconv] Fix detection of hicumL2v2p31
Qucsconv was assigning hicumL2V2p24 for the following 2.31 device
models:
  .MODEL  HICUM  NPN (LEVEL = 2  VERSION = 2.31)
  .MODEL  HICUM  NPN (LEVEL = 2  VERSION = 2.31N)

This patch allows the above VERSION values to map to hicumL2V2p31n definition.

An error message was added to catch a BJT with unknown LEVEL and VERSION
values.

Note that VERSION is stored as a double. Hence 2.31N is equal to 2.31e-9.