Correct transient error when switching to band stop filter
When switching to a band-stop filter (from low-pass or others) an
error message appeared on the console since slotUpdateSchematic()
was not initializing the schematic file name string to use and no
assignment to it was made until slotSwitchParameters() updated
the filter type. Now a default assignment is done so no transient
error occurs anymore.