xml: Recurse variable name to get its value.
Qucs expects a numerical value as a parameter, not another parameter
name.

It is unclear how the parameter depenency should be handled in general.

In the following (from bsim6.va):
 643: parameter real LINT = 0.0;
1072: parameter real DLCIG  = LINT;

If only the default value of LINT is modified during instantiation, should
the value of DLCIG be updated?

This patch sets the default values set during the model compilation.