New matching techniques for the matching tool
Matching methods implemented:
- Single stub and double stub (balanced and not balanced)
- multistage lambda/4
- cascaded LC sections (for purely real loads)
- lambda/8 with lambda/4 transformer (complex load to real source)
All the methods above were implemented for one-port and two ports matching.

The generated schematic can optionally include the S parameter
simulation block.

Added microstrip synthesis to the matching tool

It was a added a new layout on the right of the matching tool window so as to define the substrate
properties for microstrip synthesis. In this sense, now it is possible to generate microstrip
matching networks by converting the ideal transmission lines to microstrip lines.
This feature is much the same as the microstrip synthesis available on 'qucs-filter'. Indeed, I
have literally copied the functions 'calcMicrostrip()' and 'getMicrostrip()' from there.

Added documentation of the various matching methods
to '/technical/synthesis.tex'.

It was written a schematic generator so as to minimise code
redundancy. This avoids to rewrite the same code twice (at single-
port and at two-port matching functions). In other words, all schematic
generation is done at "SchematicParser()" and all those
"Create2Port_<matching method>" are no longer needed

New function for creating schematics

It was added a function for building a schematic from the components,
wires and paintings strings.

Add Precision parameter to num2str()
to allow specifying the number of significant digits.
Default is set at 6, as it was implicitly before.