Correct section header in simulation output log
In the log.txt file, the simulation output coming from the standard
error stream was prefixed with the "Errors:" header. This has been
corrected to "Errors and Warnings:" since also warnings are sent to
standard error (by qucsator). This is also aligned with the headers
used in the Simulation Messages dialog.