Added microstrip synthesis to the matching tool
It was a added a new layout on the right of the matching tool window so as to define the substrate
properties for microstrip synthesis. In this sense, now it is possible to generate microstrip
matching networks by converting the ideal transmission lines to microstrip lines.

This feature is much the same as the microstrip synthesis available on 'qucs-filter'. Indeed, I
have literally copied the functions 'calcMicrostrip()' and 'getMicrostrip()' from there.

Minor bug fixes and code comments are also included in this commit