Fixed impedance equation in transcalc
The wave impedance equation was wrong. As given by Pozar's book (Page
113, Equation 3.86):
Z_{TE} = k * eta / beta
where k is the wave number, beta is the propagation constant and eta is
the intrinsic impedance of the waveguide. The problem was that eta was
calculated for e_r = 1 and mu_r = 1