Enable to set qucsconv via environment variable.
Qucsconv is called when a Spicefile component needs to be netlisted.
The Spice code is converted to the Qucsator format and inserted into the
netlist.

The issue is that path to Qucsconv is by default expected to be in the
same prefix as Qucs.

For testing the schematic->netlist from the command line it is
convenient to tell explicitly where Qucsconv is located.

This patch does a few things:

* Introduce the QUCSCONV environment variable to override the default
  path, assume to be the same prefix as Qucs.
  Setting QUCSDIR would work just fine, but require a make install first.
  TODO update documentation along with issue #240

* Moved the QUCSCONV to main.cpp, where it is set only once with
  a newly introduced QucsSettings.Qucsconv variable.

* Replace dialog pop-up on spicefile.cpp by a statusBar message.

Remaining issue: library components also depend on the assumption that
everything is located on the same Qucs prefix. Hence QUCSDIR migh still
be necessary in some cases, as when using command line on in-three Qucs
binary for testing purporses.