Add "What's New" dialog at application startup
A "What's New" dialog can now be show when the application is
started, showing an html text on the screen.
The dialog is shown only if the Qucs version is newer than the last
version having shown a "What's New" dialog or if the user has
explicitly unchecked its "don't show again" checkbox previously.