Added ideal taper transmission line model
It was added a general model for tapered lines. It is based on dividing
the entire transmission line into differential slots (under the
assumption that the characteristic impedance in each slot is constant).
The ABCD matrix of every slot is calculated so the overall ABCD is
trivial to obtain. Finally, the S/Y-parameters are calculated.

References:
[1] Microwave engineering. David M Pozar. John Wiley and Sons. 4th
edition. Pages 261-267

Speed-up Klopfenstein taper and ABCD matrix calculations by
implementing the simple recursive formula from Grossberg, Proc. IEEE
1968, pp.1629-1630 to compute the phi(x, A) function.

Added to 'lineardevices.tex' the documentation
related to the tapered line. It includes the fundamental formulae and a
comparison between the normalized reflection coeff. of the 4 kinds of
tapers