Added ideal taper transmission line model
It was added a general model for tapered lines. It is based on dividing
the entire transmission line into differential slots (under the
assumption that the characteristic impedance in each slot is constant).
The ABCD matrix of every slot is calculated so the overall ABCD is
trivial to obtain. Finally, the S/Y-parameters are calculated.

References:
[1] Microwave engineering. David M Pozar. John Wiley and Sons. 4th
edition. Pages 261-267

Speed-up Klopfenstein taper and ABCD matrix calculations

Implemented the simple recursive formula from Grossberg, Proc. IEEE
1968, pp.1629-1630 to compute the phi(x, A) function needed for
the Klopfenstein taper calculations.

Fixed AC simulation failure + header convention

In this commit, it was added to 'lineardevices.tex' the documentation
related to the tapered line. It includes the fundamental formulae and a
comparison between the normalized reflection coeff. of the 4 kinds of
tapers

Correct index and exponent in formulas

Add missing domain for the triangular taper impedance formula

Use TeX commands for the common mathematical functions

and add a couple of comments to the formulas

Correct \beta l in \beta L in formulas

and also the formula for the reflection coefficient of a cascade
of N lines