Tapered line model
It was added a general model for tapered lines. It is based on dividing
the entire transmission line into differential slots (under the
assumption that the characteristic impedance in each slot is constant).
The ABCD matrix of every slot is calculated so the overall ABCD is
trivial to obtain.

Finally, the S/Y-parameters are calculated.

So far, it supports exponential, linear and triangular impedance
profiles.

References:
[1] Microwave engineering. David M Pozar. John Wiley and Sons. 4th
edition. Pages 261-267