Speed-up Klopfenstein taper and ABCD matrix calculations
Implemented the simple recursive formula from Grossberg, Proc. IEEE
1968, pp.1629-1630 to compute the phi(x, A) function needed for
the Klopfenstein taper calculations.
Changed the ABCD matrix to use a constant number of sections, instead
of using constant fraction of wavelength sections, since the number
of sections needed depends mainly on the impedance change across
the section, not on its length.
Corrected the ABCD matrix to Y-parameters conversion (had a missing
negative sign). Now using the a Qucs function to set the S-matrix.