Ring resonator and cap coupled filters added to Qucs-filter
Two bandpass filter topologies were added: the side coupled ring resonator and
the capacitively coupled shunt resonators topology. In addition to this, some improvements
were made concering the user interface. Specifically, it is included a preview window so
as to let the user to have a quick image of how the filter will look like. Moreover, the topology
combobox was shortened by including a new checkbox which allows to enable/disable the microstrip implementation.
This halves the size of the aforementioned combobox.

References:

[1] Quarter Wavelength Side-Coupled Ring Resonator for Bandpass filters.
Mohd Khairul Mohd Salleh. IEEE TRANSACTIONS ON MICROWAVE THEORY AND
TECHNIQUES. VOL. 56, NO. 1, JANUARY 2008

[2] Microwave engineering. David M Pozar. John Wiley and Sons. 4th Edition. Pages 443-448

Quarter wavelength ring resonator filter

It was added a Quarter wavelength ring resonator filter to the Qucs
filter synthesis tool.

References:
[1] Quarter Wavelength Side-Coupled Ring Resonator for Bandpass filters.
Mohd Khairul Mohd Salleh. IEEE TRANSACTIONS ON MICROWAVE THEORY AND
TECHNIQUES. VOL. 56, NO. 1, JANUARY 2008

New filter topology + improvements

1) It was added the capacitively coupled shunt resonators synthesis
2) The Qucs-Filter interface was restructured in the sense of that every
gui component is placed into a (vertical/horizontal) layout box.
Moreover, it was a added a display to visualize the layout of the
filter.

New layout images added

The aspect radio of most svg images was modified

1) It was a problem with the visualization of several images. The
solution was found by changing the aspect ratio.

Fixed wrong connection

It was found a bug in the schematic generated by Qucs-Filter when using
LC notch filters (Tee topology). The last horizontal wire on the right
of the schematic was being placed wrongly when the order of the filter
was an odd number.

Added combobox for microstrip implementation

The number of available filters is growing and the 'Realization'
combobox seems to be quite saturated. In this sense, it was added a
checkbox to indicate microstrip implementation. This new feature halves
the number of entries...

Added some comments...

Updated images

Disabled microstrip implementation for the new filters