Correct search string for variables name
To extract the goals variables values from the output dataset,
Qucs writes some files in the extract/ directory defining, among other
things, the text string to search in the dataset to locate the variable.
ASCO does a SUBstring match, so the search string need also to contain
the string delimiter (a trailing space), otherwise it will match any
ither string starting with the same characters.