qucshelp: fix issue #335
Workaround hardcoded DOCDIR which is not set when --prefix is not given.
Let it gracefully fail if the html files are not found.
This should allow the windows package to be portable/relocatable.
Things are searched w.r.t the executable.

qucshelp should be deprecated/replaced by plain html, see issue #434.