Qucs power combining tool (-> develop)
This is a tool for automatically create power combiners. It includes
some popular power combining structures as well as corporate combiners.

In addition to the old branch (-> master) features, it was added the LC
equivalent of a 2-way Wilkinson combiner (both single-stage and
multistage). Moreover, it was included a new panel for diplaying the
combiner topology.