Add Ctrl-C handling for the Octave output window
The QTextEdit used to show the Octave output seems not to handle the
Ctrl-C keys when it is set read only and a global QAction is set for the
same keys combination.
An EventFilter is added to catch the Ctrl-C and call the QTextEdit copy()
method; the global QAction is called anyway before this, apparently without
side effetcs.