Qucs wideband matching tool - Initial commit
This is the first commit of a new tool for the design of wideband
matching networks. It implements GRABIM ([1]) but, in favour of
simplicity, some modifications were made wrt the original paper. The
working principle is the following:

i) Selection of a set of circuit topologies that usually work for
wideband operation.
ii) Factorial search. An hypercube is defined over a pivot. At each
iteration, every vertix is evaluated. Then, the pivot is moved to the
best vertex. In case all vertices are worse than the pivot, the
hypercube shrinks. Repeat ii) for all circuit topologies.

iii) For those topologies that achieved a good matching condition in
step ii), a local optimiser is run to refine the result. So far, it is
implemented the Nelder-Mead algorithm. I decided to avoid the
calculation of the Hessian for the local optimiser because it adds much
more programming complexity and observation seems to suggest that
derivative-free algorithms can do the job easily. Before creating this
branch, I've played around with the NLopt libraries and found that
Nelder-Mead works fine. I know,... Nelder-Mead fails as soon as the
dimension of the problem runs up (>16 or so), but keep in mind that a
network of >6 elements are very hard to tune/implement in practice.

Next steps:
i)   Refine UI.
ii)  Implement PRAXIS and SUBPLEX
iii) Extend the current testbench.

Reference:
[1] Broadband Direct-Coupled and Matching RF Networks. Thomas Cuthber.
1999