Use category and module index for all component tabs
so that every icon is univocally linked to a component.
While this is not needed for the usual components tabs, where the
components icons are placed in the same order as in their category so
the corresponding component can be found looking at the icon position
in the tab it's useful to unify the handling of icons with the other
cases, whey they are placed in a different order, as for
the "Search results" tab.
This will also help if other tabs with icons placed in an order not
reflecting the component categories order will be used, as for
example a "Recently used" tab.