Move all symbol definition stuff to SymbolWidget
The symbol definition for standard components is now handled by
SymbolWidget() and not by QucsLib(), so this latter does not need to
know how to handle the standard components (which do not define a
symbol).
Handling of components which do not define a symbol but should use
the library default symbol was moved to qucslib_common.h previously
so also this case does not need to be explicitly handled by QucsLib
anymore.