vhdl : fix char* to QString conversion.
The t.toLatin1() creates a QByteArray whose lifetime ends at the end of the
expression. But &p is now a pointing to someplace whose existence is no
guaranteed after the line is executed.
To avoid this, we can do it in two steps, storing the QByteArray first and
doing the conversion to double afterwards, like

(Thank you in3otd for the detective work in the pull request #608.

This is causing VHDL simulation to fail as such:
  0.0.18 : netCLEAR <= '1'; wait for 10 ns;
  before fix: netCLEAR <= '1'; wait for 10 1;

The `Verilog_Time()` is not using `&p` afterwards, modified for
consistency.