Refactor setter getter for colored buttons.
Due to platform differences creating a simple color picker button seems
difficult.
See discussion in issue #74.

This patch used part of PR #135 to implement the button with a pixmap on
it. Stylesheets could be used instead.

This patch also pulls all the get/set of colors into buttons into misc::