Delta markers using QMap to handle data
In this commit, the way the marker information (value, color and other
properties) was changed so as to avoid Marker.dat files which polluted
the file system and involved an unnecessary number of I/O file
operations.

In this sense, the Marker.dat table was replaced by a QMap structure in
'Diagram' class. This way, each diagram contains the necessary info of
its markers and it is updated when calling the paint() event. Moreover,
this table is passed to every marker in order to let 'markerdialog' know
what markers can be used as reference markers.

Regarding backwards compatibility, this code contains a legacy function
for loading the old marker format from the .dpl file.

Finally, it was added an extra (dotted) line to the delta markers so as
to give a visual clue about where the reference marker is.