Addition of the Ohmmeter Module
This probe should perhaps be called "Impedance Meter" instead, since
it measures the impedance, with its angular component and not only
the resistance, I kept it for simplicity and so it would fit on the
menus but maybe something to look into in the future.

The working loop is farly simple, it will act as a voltmeter on a first
instance, if it doesn't detect any other source it will then act as an
ohmmeter. By setting a current of 1A on its branch, and by following ohm's
law, it will then read the voltage differential between its two nodes
since if V=1*R,V=R, thus determining the impedance.

Hoping this doesn't update the previous commit and instead adds to the
already open Pull Request. Updated the Wprobe.png and its cpp in an
attempt to streamline it with the other components. Will add a simple
pdf with working examples of the ohmmeter on a comment.
Thanks for the attention.