Addition of the Internal Resistance Parameter
Added the option of adding an internal resistance to both voltage sources,
ac and dc, which would equal a resistor in series with them.
The same applies to Iprobe. The default values for these components are 0 Ohms.
In the voltmeter, another series resistance is applied, but the default value is 1e12, I can change this if another value is prefered. The Wattmeter also has internal resistance, both for the voltage and the current part, the same principals were applied, as they were to their Probes.
Regarding the current sources, the internal resistor will result in a paralel with the component, and the default value was set to 1e12ohms as well.
I will add some working examples in a pdf.
Thanks to andresmmera for all his help and patience on this topic.
Edit:Fixed the default values of some components
Edit2:Added message, an Ri of 0 disables the parameter