Misc GUI fixes
* Added ccoupled_shunt_resonators.h to Makefile.am as suggested by
ra3xdh

* ccoupled_shunt_resonators.cpp, qw_coupled_ring_filter.cpp: Double
format lengths converted to string using num2str()

* lc_filter.cpp: The last component of a LCL filter (even order only)
was not connected

* qf_filter.cpp: It was the 'dBS11' and 'dBS21' variables were renamed
so as to be in line with the other filter implementations

*qucsfilter.cpp: As suggested by ra3xdh, the behaviour of the labels in
the quarter wave side coupled resonator bandpass filter was fixed.
Moreover, it was added an icon for the RFEDD based filters (which was
missing before).

On the other hand, the logic of the microstrip panel was fixed in order
to avoid the undefined behaviour described by ra3xdh at the Github
thread.

Last but not least, QucsFilter::calculateFilter() was modified so as to
avoid redundant error messages when even order Chebyshev filters are
selected. In addition to this, the error message was changed, so far it
said "Even order Chebyshev can't be realized with passive filters."
Actually, it is possible to implement passive filters of even order
having RS != RL