Fixes and improvements to the commited modules.
The aspect ratio of the two new diagrams is now changeable.
Added some more clarity to the internal resistance parameter.
Deleted a lot of unused variables and functions.
Changed names of the diagrams and one variable to better follow convention.
Added components and diagrams to CMakelists.
Most of this was achieved under the guidance on Andrés Mera, I thank him for
all his help.