Solved missing reference marker issues
This commit fixes the issues described by in3otd
(https://github.com/Qucs/qucs/pull/407#issuecomment-312516053)

The fix consist on updating the ActiveMarkers map in the paint() event
so as to know if the reference marker has been removed or if its name
has changed. In these cases, the marker pointing to that reference is
set to be a convenitional marker.

In the same line, if the user activates the delta mode for the reference
marker, those markers pointing to that reference are converted into
conventional markers