Check if variables were changed before asking if the user wants to update them
It was noticed that every tie the user triggers the close event the
application is asking to update the values even if the user did nothing.
In ordet to fix this, a check was added before throwing the message box
asking for that.

Moreover, this commit includes a small fix for handling component
properties without unit