Change the time resolution for the update-graphs? condition
It was observed that when moving the tuning slider fast the results
didn't change according to it. After some research, it was found that
the condition for updating graphs had 1 second resolution. That's fine
when Qucs works in "conventional mode" since it happens 5 seconds or so
until the user hits the simulation button again. However, in tuning
mode, the user may run consecutive simulations whose length < 1 s (for
example, simple DC analysis...)...

In order to tackle this, now it is used resultion of 1ms