Replaced QLineedit by QComboBox for freq input at phasordiagram dialog
In this commit it was replaced the QLineEdit for the frequency input at
the diagram dialog of the phasor diagram by a QComboBox showing only the
available frequencies.

In order to do this, a function named 'LoadAvailableFreqs()' was added
to diagramdialog constructor. The purpose of this function is to look
for the content of the 'acfrequency' field the current dataset and
returns that data as a QStringList which feeds the QComboBox.
'LoadAvailableFreqs()' is called at the constructor but also when the
click() event of the 'ChooseData' combo is triggered.