Correct marker misbehaviour when pointing to hidden data
When moving a marker back and forth, it points to (0,0) if the graph
contains data that lies outside the diagram limits. This commit solves
that issue by checking (only for Cartesian diagrams) if the y-value
exceeds the diagram limints at Marker::createText().

Additionally, a big red point is displayed so as to indicate that the
point is not being displayed