Correct marker behaviour when pointing outside diagram
for the rectangular diagram.
If a graph contained data that were outside the diagram limits, the
marker was forced to point to (0,0). This commit solves
that issue by checking if the y-value exceeds the diagram limits
and setting a flag in the Marker object.
(most of the code comes from Andres)
    Additionally, a big red point is displayed so as to indicate that the
    point is not being displayed