Removed phasordiagram and waveac from the build system
The aim of this commit is to lay aside the waveac and phasordiagram
code. This code is difficultinh the merge of old PRs and entagles
(even more) the Qucs code. In this sense, phasordiagram.h/.cpp and
waveac.h/.cpp were removed from "CMakeLists.txt" and "Makefile.am". The
most relevant code spread over diagram.cpp/h, marker.cpp/h, etc. was
commented whereas those portions of code that intrude
existing functions were removed since it'll be needed to figure out an
alternative non-intrusive way to implement that feature.