Added function for copying the examples in the user workspace
Since the examples are located in /usr/share/ it is not possible to run
simulations nor save changes in the sample schematics. In order to
tackle this problem, it was added a new menu entry, 'File/Copy Examples
in the home directory' to copy the whole examples directory in the home
dir. Moreover, it was added an informative pop-up window the
'File/Examples' so as to warn the  user against the read-only nature of
/usr/share/ (at least in unix-like systems... I think we need to check
the OS in upcoming commits...)

The directory copy function was taken from
https://stackoverflow.com/questions/2536524/copy-directory-using-qt

Copy examples to the home dir

The reason for this new feature is that the examples are located at /usr/share, which is a read-only directory (so the user cannot run simulations nor save changes).

Remove stuff related to the lock feature