Create checkbox under App Settings for the description feature
In this commit, the "show schematic description while hovering mouser"
feature is set as a feature. Since this feature, should be common for
all the schematic files in the project tree, I considered the most
appropiate place to put this option is under the Application Settings
menu.

The menu is unchecked by default.