Use pkg-config to find Qt4
The exceedingly complex existing ad-hoc solution does not find Qt4 on current
Gentoo Linux. It is left as a fallback if pkg-config should fail for some
reason.

Variables QT_LDFLAGS and QT_INCLUDES have been renamed QT_LIBS and QT_CFLAGS
to match autoconf convention.

In order to better show the actual changes I have not changed the indentation,
that's done in the following commit.