New UI option to add a SP simulation box in qucsattenuator
It was added a checkbox in the qucsatenuattor UI so as to let the user
to attach a SP simulation component to the attenuator circuit. Since
resistive attenuators are wideband circuit (yes, in practice they're
tight to the packaging parasitics...), it doesn't ask for the frequency
sweep. In this sense, the SP sweep was fixed to 50MHz-3GHz... the user
may modify that range in the Qucs schematics later...