CMake Tests: Extensive overhaul
This commit solves the issue with various files loaded
in the same API tests, potentially contaminating the environment
when debugging.

It also adds a flag which is by default off, allowing the build
of those many binaries to be turned off and on.

In addition, the check target has been extended and now also
loads Verilog Schematic files, dumping them back out into Verilog
Schematics as well as the Legacy Schematic format.